+++ Realm/RLMObjectStore.mm
@@ -28,49 +28,34 @@
 #import <objc/runtime.h>
 
 static void RLMVerifyAndAlignColumns(RLMObjectSchema *tableSchema, RLMObjectSchema *objectSchema) {
-    // FIXME - this method should calculate all mismatched columns, and missing/extra columns, and include
-    //         all of this information in a single exception
-    // FIXME - verify property attributes
-
-    // check count
-    if (tableSchema.properties.count != objectSchema.properties.count) {
-        @throw [NSException exceptionWithName:@"RLMException"
-                                       reason:@"Column count does not match interface - migration required"
-                                     userInfo:nil];
-    }
-
     NSMutableArray *properties = [NSMutableArray arrayWithCapacity:objectSchema.properties.count];
+    NSMutableArray *exceptionMessages = [NSMutableArray array];
 
     // check to see if properties are the same
     for (RLMProperty *tableProp in tableSchema.properties) {
         RLMProperty *schemaProp = objectSchema[tableProp.name];
-        if (![tableProp.name isEqualToString:schemaProp.name]) {
-            @throw [NSException exceptionWithName:@"RLMException"
-                                           reason:@"Existing property does not match interface - migration required"
-                                         userInfo:@{@"property num": @(tableProp.column),
-                                                    @"existing property name": tableProp.name,
-                                                    @"new property name": schemaProp.name}];
+        if (!schemaProp) {
+            [exceptionMessages addObject:[NSString stringWithFormat:@"Property '%@' is missing from latest object model.", tableProp.name]];
+            continue;
         }
         if (tableProp.type != schemaProp.type) {
-            @throw [NSException exceptionWithName:@"RLMException"
-                                           reason:@"Property types do not match - migration required"
-                                         userInfo:@{@"property name": tableProp.name,
-                                                    @"existing property type": RLMTypeToString(tableProp.type),
-                                                    @"new property type": RLMTypeToString(schemaProp.type)}];
+            [exceptionMessages addObject:[NSString stringWithFormat:@"Property types for '%@' property do not match. Old type '%@', new type '%@'.",
+                                          tableProp.name, RLMTypeToString(tableProp.type), RLMTypeToString(schemaProp.type)]];
+            continue;
         }
         if (tableProp.type == RLMPropertyTypeObject || tableProp.type == RLMPropertyTypeArray) {
             if (![tableProp.objectClassName isEqualToString:schemaProp.objectClassName]) {
-                @throw [NSException exceptionWithName:@"RLMException"
-                                               reason:@"Property objectClass does not match - migration required"
-                                             userInfo:@{@"property name": tableProp.name,
-                                                        @"existign objectClass": tableProp.objectClassName,
-                                                        @"new property name": schemaProp.objectClassName}];
+                [exceptionMessages addObject:[NSString stringWithFormat:@"Target object type for property '%@' does not match. Old type '%@', new type '%@'.",
+                                              tableProp.name, tableProp.objectClassName, schemaProp.objectClassName]];
             }
         }
         if (tableProp.isPrimary != schemaProp.isPrimary) {
-            @throw [NSException exceptionWithName:@"RLMException"
-                                           reason:@"Property primary key designation does not match - migration required"
-                                         userInfo:@{@"property name": tableProp.name}];
+            if (tableProp.isPrimary) {
+                [exceptionMessages addObject:[NSString stringWithFormat:@"Property '%@' is no longer a primary key.", tableProp.name]];
+            }
+            else {
+                [exceptionMessages addObject:[NSString stringWithFormat:@"Property '%@' has been made a primary key.", tableProp.name]];
+            }
         }
 
         // align
@@ -78,6 +63,20 @@ static void RLMVerifyAndAlignColumns(RLMObjectSchema *tableSchema, RLMObjectSche
         [properties addObject:schemaProp];
     }
 
+    // check for new missing properties
+    for (RLMProperty *schemaProp in objectSchema.properties) {
+        if (!tableSchema[schemaProp.name]) {
+            [exceptionMessages addObject:[NSString stringWithFormat:@"Property '%@' has been added to latest object model.", schemaProp.name]];
+        }
+    }
+
+    // throw if errors
+    if (exceptionMessages.count) {
+        @throw [NSException exceptionWithName:@"RLMException"
+                                       reason:[NSString stringWithFormat:@"Migration is required for object type '%@' due to the following errors:\n- %@",
+                                               objectSchema.className, [exceptionMessages componentsJoinedByString:@"\n- "]]
+                                     userInfo:nil];
+    }
     // ensure the order of the properties matches the column order in the table
     objectSchema.properties = properties;
 }
@@ -108,13 +107,7 @@ static void RLMCreateColumn(RLMRealm *realm, tightdb::Table &table, RLMProperty 
     }
 }
 
-void RLMRealmInitializeReadOnlyWithSchema(RLMRealm *realm, RLMSchema *targetSchema) {
-    if (RLMRealmSchemaVersion(realm) == RLMNotVersioned) {
-        @throw [NSException exceptionWithName:@"RLMException"
-                                       reason:@"Cannot open an uninitialized realm in read-only mode"
-                                     userInfo:nil];
-    }
-
+void RLMRealmSetSchema(RLMRealm *realm, RLMSchema *targetSchema, bool verify) {
     realm.schema = [targetSchema copy];
 
     for (RLMObjectSchema *objectSchema in realm.schema.objectSchema) {
@@ -122,83 +115,69 @@ void RLMRealmInitializeReadOnlyWithSchema(RLMRealm *realm, RLMSchema *targetSche
         objectSchema->_table = RLMTableForObjectClass(realm, objectSchema.className);
         if (objectSchema->_table) {
             RLMObjectSchema *tableSchema = [RLMObjectSchema schemaFromTableForClassName:objectSchema.className realm:realm];
-            RLMVerifyAndAlignColumns(tableSchema, objectSchema);
+            if (verify) {
+                RLMVerifyAndAlignColumns(tableSchema, objectSchema);
+            }
         }
         objectSchema.accessorClass = RLMAccessorClassForObjectClass(objectSchema.objectClass, objectSchema);
     }
 }
 
-void RLMRealmInitializeWithSchema(RLMRealm *realm, RLMSchema *targetSchema) {
-    [realm beginWriteTransaction];
-
-    @try {
-        // check to see if this is the first time loading this realm
-        bool firstInitialization = RLMRealmSchemaVersion(realm) == RLMNotVersioned;
-        if (firstInitialization) {
-            // set initial version
-            RLMRealmSetSchemaVersion(realm, 0);
-        }
-
-        // set the schema, mutating if we are initializing the db for the first time
-        RLMRealmSetSchema(realm, targetSchema, firstInitialization);
-    }
-    @finally {
-        // FIXME: should rollback on exceptions rather than commit once that's implemented
-        [realm commitWriteTransaction];
-    }
-}
-
-bool RLMRealmSetSchema(RLMRealm *realm, RLMSchema *targetSchema, bool initializeSchema) {
+void RLMRealmCreateTables(RLMRealm *realm, RLMSchema *targetSchema, bool updateExisting) {
     realm.schema = [targetSchema copy];
 
+    // first pass to create missing tables
     bool changed = false;
-    if (initializeSchema) {
-        // first pass to create missing tables
-        for (RLMObjectSchema *objectSchema in realm.schema.objectSchema) {
-            bool created = false;
-            objectSchema->_table = RLMTableForObjectClass(realm, objectSchema.className, created);
-            changed |= created;
+    NSMutableArray *objectSchemaToUpdate = [NSMutableArray array];
+    for (RLMObjectSchema *objectSchema in realm.schema.objectSchema) {
+        bool created = false;
+        objectSchema->_table = RLMTableForObjectClass(realm, objectSchema.className, created);
+        changed |= created;
+
+        // we will modify tables for any new objectSchema (table was created) or for all if updateExisting is true
+        if (updateExisting || created) {
+            [objectSchemaToUpdate addObject:objectSchema];
         }
+    }
 
-        // second pass adds/removes columns appropriately
-        for (RLMObjectSchema *objectSchema in realm.schema.objectSchema) {
-            RLMObjectSchema *tableSchema = [RLMObjectSchema schemaFromTableForClassName:objectSchema.className realm:realm];
+    // second pass adds/removes columns for objectSchemaToUpdate
+    for (RLMObjectSchema *objectSchema in objectSchemaToUpdate) {
+        RLMObjectSchema *tableSchema = [RLMObjectSchema schemaFromTableForClassName:objectSchema.className realm:realm];
 
-            // add missing columns
-            for (RLMProperty *prop in objectSchema.properties) {
-                // add any new properties (new name or different type)
-                if (!tableSchema[prop.name] || ![prop isEqualToProperty:tableSchema[prop.name]]) {
-                    RLMCreateColumn(realm, *objectSchema->_table, prop);
-                    changed = true;
-                }
-            }
-
-            // remove extra columns
-            for (int i = (int)tableSchema.properties.count - 1; i >= 0; i--) {
-                RLMProperty *prop = tableSchema.properties[i];
-                if (!objectSchema[prop.name] || ![prop isEqualToProperty:objectSchema[prop.name]]) {
-                    objectSchema->_table->remove_column(prop.column);
-                    changed = true;
-                }
-            }
-
-            // update table metadata
-            if (objectSchema.primaryKeyProperty != nil) {
-                // if there is a primary key set, check if it is the same as the old key
-                if (tableSchema.primaryKeyProperty == nil || ![tableSchema.primaryKeyProperty isEqual:objectSchema.primaryKeyProperty]) {
-                    RLMRealmSetPrimaryKeyForObjectClass(realm, objectSchema.className, objectSchema.primaryKeyProperty.name);
-                    changed = true;
-                }
-            }
-            else if (tableSchema.primaryKeyProperty) {
-                // there is no primary key, so if thre was one nil out
-                RLMRealmSetPrimaryKeyForObjectClass(realm, objectSchema.objectClass, nil);
+        // add missing columns
+        for (RLMProperty *prop in objectSchema.properties) {
+            // add any new properties (new name or different type)
+            if (!tableSchema[prop.name] || ![prop isEqualToProperty:tableSchema[prop.name]]) {
+                RLMCreateColumn(realm, *objectSchema->_table, prop);
                 changed = true;
             }
         }
 
-        // FIXME - remove deleted tables
+        // remove extra columns
+        for (int i = (int)tableSchema.properties.count - 1; i >= 0; i--) {
+            RLMProperty *prop = tableSchema.properties[i];
+            if (!objectSchema[prop.name] || ![prop isEqualToProperty:objectSchema[prop.name]]) {
+                objectSchema->_table->remove_column(prop.column);
+                changed = true;
+            }
+        }
+
+        // update table metadata
+        if (objectSchema.primaryKeyProperty != nil) {
+            // if there is a primary key set, check if it is the same as the old key
+            if (tableSchema.primaryKeyProperty == nil || ![tableSchema.primaryKeyProperty isEqual:objectSchema.primaryKeyProperty]) {
+                RLMRealmSetPrimaryKeyForObjectClass(realm, objectSchema.className, objectSchema.primaryKeyProperty.name);
+                changed = true;
+            }
+        }
+        else if (tableSchema.primaryKeyProperty) {
+            // there is no primary key, so if thre was one nil out
+            RLMRealmSetPrimaryKeyForObjectClass(realm, objectSchema.objectClass, nil);
+            changed = true;
+        }
     }
+
+    // FIXME - remove deleted tables
 
     for (RLMObjectSchema *objectSchema in realm.schema.objectSchema) {
         // cache table instances on objectSchema
@@ -213,8 +192,6 @@ bool RLMRealmSetSchema(RLMRealm *realm, RLMSchema *targetSchema, bool initialize
         //         us to have multiple accessors for each type/instance combination
         objectSchema.accessorClass = RLMAccessorClassForObjectClass(objectSchema.objectClass, objectSchema);
     }
-
-    return changed;
 }
 
 static inline void RLMVerifyInWriteTransaction(RLMRealm *realm) {